95% done, need to compelte real time update on sessiond eleteion

+++ backend/api/sessions.cjs
@@ -2,6 +2,8 @@ const express = require('express')
 const sessionRouter = express.Router();
 const myPrisma = require('../db/myPrisma.cjs');
 const prisma = require('../db/connection.cjs')
+const { verifyToken } = require('../auth/middleware.cjs');
+
 
 //get all the sessions
 sessionRouter.get("/", async (req, res, next) => {
@@ -70,19 +72,36 @@ sessionRouter.patch('/:id', async (req, res, next) => {
 
 
 // Delete session by ID
-sessionRouter.delete("/:id", async (req, res, next) => {
+sessionRouter.delete("/:id", verifyToken, async (req, res, next) => {
   const { id } = req.params;
+  const { uid } = req.user;
 
   try {
-    await prisma.reservation.deleteMany({
-      where: { session_id: Number(id) },
+    const user = await prisma.user.findUnique({
+      where: {
+        uid: uid,
+      },
     });
+
+    if (!user) {
+      return res.status(404).send({ error: "User not found" });
+    }
+
+    // await prisma.reservation.deleteMany({
+    //   where: { session_id: Number(id) },
+    // });
+
+    // await prisma.availability.deleteMany({
+    //   where: { session_id: Number(id) },
+    // });
 
     const deletedSession = await prisma.session.delete({
       where: { id: Number(id) },
     });
+
     res.send(deletedSession);
   } catch (error) {
+    console.error("Error deleting session:", error);
     next(error);
   }
 });
